--- .gitattributes
@@ -1,1 +1,0 @@
-demo/**/* linguist-vendored
