+++ components/sui.module.ts
@@ -1,0 +1,32 @@
+import {NgModule} from "@angular/core";
+
+import {SuiAccordionModule} from "./accordion/accordion.module";
+import {SuiCheckboxModule} from "./checkbox/checkbox.module";
+import {SuiCollapseModule} from "./collapse/collapse.module";
+import {SuiDimmerModule} from "./dimmer/dimmer.module";
+import {SuiDropdownModule} from "./dropdown/dropdown.module";
+import {SuiMessageModule} from "./message/message.module";
+import {SuiProgressModule} from "./progress/progress.module";
+import {SuiRatingModule} from "./rating/rating.module";
+import {SuiSearchModule} from "./search/search.module";
+import {SuiTabsModule} from "./tabs/tab.module";
+import {SuiSelectModule} from "./select/select.module";
+import {SuiTransitionModule} from "./transition/transition.module";
+
+@NgModule({
+    exports: [
+        SuiAccordionModule,
+        SuiCheckboxModule,
+        SuiCollapseModule,
+        SuiDimmerModule,
+        SuiDropdownModule,
+        SuiMessageModule,
+        SuiProgressModule,
+        SuiRatingModule,
+        SuiSearchModule,
+        SuiSelectModule,
+        SuiTabsModule,
+        SuiTransitionModule
+    ]
+})
+export class SuiModule {}
