+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "ng2-semantic-ui",
-  "main": "ng2-semantic-ui.js",
-  "typings": "ng2-semantic-ui.d.ts",
+  "main": "index.js",
+  "typings": "index.d.ts",
   "version": "0.4.7",
   "description": "Angular 2 Semantic UI Components",
   "repository": {
